+++ internal/app/wwctl/node/imprt/main.go
@@ -12,8 +12,7 @@ import (
 	"github.com/warewulf/warewulf/internal/pkg/api/routes/wwapiv1"
 	apiutil "github.com/warewulf/warewulf/internal/pkg/api/util"
 	"github.com/warewulf/warewulf/internal/pkg/node"
-	"github.com/warewulf/warewulf/internal/pkg/wwlog"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 func CobraRunE(cmd *cobra.Command, args []string) error {
@@ -66,10 +65,6 @@ func CobraRunE(cmd *cobra.Command, args []string) error {
 				}
 				if importMap[line[0]] == nil {
 					importMap[line[0]] = new(node.NodeConf)
-				}
-				ok := importMap[line[0]].SetLopt(records[0][j], line[j])
-				if !(ok) {
-					wwlog.Debug("Could not import %s\n", line[j])
 				}
 			}
 		}