+++ internal/pkg/warewulfd/provision_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	warewulfconf "github.com/warewulf/warewulf/internal/pkg/config"
-	"github.com/warewulf/warewulf/internal/pkg/node"
+	"github.com/warewulf/warewulf/internal/pkg/testenv"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
 )
 
@@ -35,11 +35,9 @@ var provisionSendTests = []struct {
 }
 
 func Test_ProvisionSend(t *testing.T) {
-	conf_file, err := os.CreateTemp(os.TempDir(), "ww-test-nodes.conf-*")
-	assert.NoError(t, err)
-	defer conf_file.Close()
-	{
-		_, err := conf_file.WriteString(`WW_INTERNAL: 45
+	env := testenv.New(t)
+
+	env.WriteFile(t, "etc/warewulf/nodes.conf", `WW_INTERNAL: 45
 nodeprofiles:
   default:
     container name: suse
@@ -48,6 +46,8 @@ nodes:
     network devices:
       default:
         hwaddr: 00:00:00:ff:ff:ff
+    profiles:
+    - default
   n2:
     network devices:
       default:
@@ -61,25 +61,13 @@ nodes:
     ipxe template: test
     kernel:
       override: 1.1.1`)
-		assert.NoError(t, err)
-	}
-	assert.NoError(t, conf_file.Sync())
-	node.ConfigFile = conf_file.Name()
-
 	// create a  arp file as for grub we look up the ip address through the arp cache
-	arp_file, err := os.CreateTemp(os.TempDir(), "ww-arp")
-	assert.NoError(t, err)
-	defer arp_file.Close()
-	{
-		_, err := arp_file.WriteString(`IP address       HW type     Flags       HW address            Mask     Device
+
+	env.WriteFile(t, "arpcache", `IP address       HW type     Flags       HW address            Mask     Device
 10.10.10.10    0x1         0x2         00:00:00:ff:ff:ff     *        dummy
 10.10.10.11    0x1         0x2         00:00:00:00:ff:ff     *        dummy
 10.10.10.12    0x1         0x2         00:00:00:00:00:ff     *        dummy`)
-		assert.NoError(t, err)
-	}
-	assert.NoError(t, arp_file.Sync())
-	SetArpFile(arp_file.Name())
-
+	SetArpFile(path.Join(env.BaseDir, "arpcache"))
 	conf := warewulfconf.Get()
 	containerDir, imageDirErr := os.MkdirTemp(os.TempDir(), "ww-test-container-*")
 	assert.NoError(t, imageDirErr)
@@ -114,17 +102,13 @@ nodes:
 	dbErr := LoadNodeDB()
 	assert.NoError(t, dbErr)
 
-	provisionDir, provisionDirErr := os.MkdirTemp(os.TempDir(), "ww-test-provision-*")
-	assert.NoError(t, provisionDirErr)
-	defer os.RemoveAll(provisionDir)
-	conf.Paths.WWProvisiondir = provisionDir
 	conf.Warewulf.Secure = false
-	wwlog.SetLogLevel(wwlog.DEBUG)
-	assert.NoError(t, os.MkdirAll(path.Join(provisionDir, "overlays", "n1"), 0700))
-	assert.NoError(t, os.WriteFile(path.Join(provisionDir, "overlays", "n1", "__SYSTEM__.img"), []byte("system overlay"), 0600))
-	assert.NoError(t, os.WriteFile(path.Join(provisionDir, "overlays", "n1", "__RUNTIME__.img"), []byte("runtime overlay"), 0600))
-	assert.NoError(t, os.WriteFile(path.Join(provisionDir, "overlays", "n1", "o1.img"), []byte("specific overlay"), 0600))
+	assert.NoError(t, os.MkdirAll(path.Join(conf.Paths.WWProvisiondir, "overlays", "n1"), 0700))
+	assert.NoError(t, os.WriteFile(path.Join(conf.Paths.WWProvisiondir, "overlays", "n1", "__SYSTEM__.img"), []byte("system overlay"), 0600))
+	assert.NoError(t, os.WriteFile(path.Join(conf.Paths.WWProvisiondir, "overlays", "n1", "__RUNTIME__.img"), []byte("runtime overlay"), 0600))
+	assert.NoError(t, os.WriteFile(path.Join(conf.Paths.WWProvisiondir, "overlays", "n1", "o1.img"), []byte("specific overlay"), 0600))
 
+	wwlog.SetLogLevel(wwlog.DEBUG)
 	for _, tt := range provisionSendTests {
 		t.Run(tt.description, func(t *testing.T) {
 			req := httptest.NewRequest(http.MethodGet, tt.url, nil)
@@ -136,7 +120,9 @@ nodes:
 
 			data, readErr := io.ReadAll(res.Body)
 			assert.NoError(t, readErr)
-			assert.Equal(t, tt.body, string(data))
+			if tt.body != "" {
+				assert.Equal(t, tt.body, string(data))
+			}
 			assert.Equal(t, tt.status, res.StatusCode)
 		})
 	}